Create README.md
This is the third portfolio app for the Udacity iOS Developer Nanodegree. 
App with a map that shows information posted by other students. The map contains pins that show the location where other students have reported studying. By tapping on the pin users can see a URL for something the student finds interesting. The user is able to add their own data by posting a string that can be reverse geocoded to a location, and a URL.